refactor(ui): extract SystemPage pure data into SystemPage.constants
Move formatUptime, CATEGORY_COLORS, and CSV_TABLES out of SystemPage.tsx
into a co-located SystemPage.constants module. No behavior change — pure
data/helpers with no React dependency.
